+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## next
+
+- Enhance the `text` option for defining the text content of the column headers.
+
 ## 1.1.0 (Dec 15, 2018)
 
 - Add `headers` option for customizing the column headers (#20).